Add metadata (including a license) to the package
In particular, add the MPL 2.0 license.
This deviates from RGBDS' current license (Expat / "MIT"), knowingly so.

RGBDS has had a troubled history regarding licensing
(see https://github.com/gbdev/rgbds/issues/128), but ultimately, it was
cleanly(?) licensed under the Expat/"MIT" license.

However, it seems correct to license "rsgbds" under a *different* license,
as it is entirely (at this point) my [@ISSOtm's] code.
It borrows quite heavily from the original structure of RGBDS, but most of
that has become my own and @Rangi42's code in the past few years.
(For example, we can claim with certainty the entirety of RGBLINK, RGBFIX,
and RGBGFX as our own, likewise the entirety of RGBASM's lexer.)
Rangi has OK'd this license change as well.

The choice of the MPL 2.0 boils down to being in favour of copyleft, but
also understanding that a GPL3'd RGBDS would be a big barrier to its
adoption: an assembly toolchain is a prime candidate for integrating into
compiler toolchains (and we are, in fact, aiming for that); then those
toolchains are likely to be embedded further.
I like what the GPL strives to do, but many people take issue at its
virality, and we can do without that barrier to entry.